+++ CHANGES.md
@@ -6,6 +6,8 @@ Changelog for Hatch Gettext
 
  - Add options to use `intltool` to regenerate the .pot template and check for 
    left out files.
+ - If LINGUAS file or environment variable is specified, only generate mo files
+   for languages it specifies.
 
 0.6.0 (2024-03-27)
 ------------------